added xib file and made redisign
