+++ Themoviedb/CellMoviesXIB.xib
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="148" id="KGk-i7-Jjw" customClass="CellMoviesXIB" customModule="Themoviedb" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="527" height="148"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="527" height="147.5"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Image" translatesAutoresizingMaskIntoConstraints="NO" id="YMD-W3-Hpw">
+                        <rect key="frame" x="16" y="8" width="75" height="113"/>
+                        <constraints>
+                            <constraint firstAttribute="width" constant="75" id="QgQ-Fw-R7F"/>
+                            <constraint firstAttribute="height" constant="113" id="hXC-fo-sns"/>
+                        </constraints>
+                    </imageView>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="LabelTitle" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="LUo-af-1Z0">
+                        <rect key="frame" x="99" y="15" width="368" height="19"/>
+                        <fontDescription key="fontDescription" type="system" weight="semibold" pointSize="16"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="LabelSubTitle" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="TcT-3W-NbV">
+                        <rect key="frame" x="99" y="38" width="368" height="19.5"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="16"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Image-1" translatesAutoresizingMaskIntoConstraints="NO" id="FeG-ir-b7m">
+                        <rect key="frame" x="483" y="113" width="8" height="13"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="13" id="4UU-qX-yWz"/>
+                            <constraint firstAttribute="width" constant="8" id="lnL-WA-qh2"/>
+                        </constraints>
+                    </imageView>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="FeG-ir-b7m" firstAttribute="leading" secondItem="LUo-af-1Z0" secondAttribute="trailing" constant="16" id="1sU-LM-GEC"/>
+                    <constraint firstItem="FeG-ir-b7m" firstAttribute="leading" secondItem="TcT-3W-NbV" secondAttribute="trailing" constant="16" id="ETM-YP-B09"/>
+                    <constraint firstAttribute="bottom" secondItem="YMD-W3-Hpw" secondAttribute="bottom" constant="8" id="SAA-a7-21u"/>
+                    <constraint firstItem="YMD-W3-Hpw" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="16" id="aYo-kA-gN9"/>
+                    <constraint firstItem="FeG-ir-b7m" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="topMargin" constant="58" id="ahI-yx-4TV"/>
+                    <constraint firstItem="LUo-af-1Z0" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="15" id="fku-2k-aHk"/>
+                    <constraint firstItem="YMD-W3-Hpw" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="8" id="hO7-5E-EjF"/>
+                    <constraint firstAttribute="bottom" secondItem="TcT-3W-NbV" secondAttribute="bottom" constant="15" id="iAK-Nz-fWC"/>
+                    <constraint firstItem="TcT-3W-NbV" firstAttribute="top" secondItem="LUo-af-1Z0" secondAttribute="bottom" constant="4" id="mmm-dB-dM7"/>
+                    <constraint firstItem="TcT-3W-NbV" firstAttribute="leading" secondItem="YMD-W3-Hpw" secondAttribute="trailing" constant="8" id="mnd-gh-beQ"/>
+                    <constraint firstItem="LUo-af-1Z0" firstAttribute="leading" secondItem="YMD-W3-Hpw" secondAttribute="trailing" constant="8" id="nQz-PT-O26"/>
+                    <constraint firstAttribute="trailingMargin" secondItem="FeG-ir-b7m" secondAttribute="trailing" constant="16" id="wlO-dF-ku5"/>
+                    <constraint firstAttribute="bottomMargin" secondItem="FeG-ir-b7m" secondAttribute="bottom" constant="58" id="zvA-TJ-DYg"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="imageMovies" destination="YMD-W3-Hpw" id="Mhs-ja-8s3"/>
+                <outlet property="labelDescriptions" destination="TcT-3W-NbV" id="yQd-0H-1q2"/>
+                <outlet property="labelTitle" destination="LUo-af-1Z0" id="qgM-GV-Ifc"/>
+                <outlet property="showImage" destination="FeG-ir-b7m" id="Mcv-ax-4vE"/>
+            </connections>
+            <point key="canvasLocation" x="287.68115942028987" y="76.339285714285708"/>
+        </tableViewCell>
+    </objects>
+    <resources>
+        <image name="Image" width="75" height="113"/>
+        <image name="Image-1" width="8" height="13"/>
+    </resources>
+</document>
