--- Themoviedb/SearchMoviesTableViewCell.swift
@@ -1,24 +1,0 @@
-//
-//  SearchMoviesTableViewCell.swift
-//  Themoviedb
-//
-//  Created by Кирилл Кибешев on 16/07/2019.
-//  Copyright © 2019 Кирилл Кибешев. All rights reserved.
-//
-
-import UIKit
-
-class SearchMoviesTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
