+++ Themoviedb/SearchMoviesViewController.swift
@@ -33,6 +33,7 @@ class SearchMoviesViewController: UIViewController {
         tableView?.delegate = self
         tableView?.dataSource = self
         tableView?.tableFooterView = UIView()
+        tableView?.register(UINib(nibName: "CellMoviesXIB", bundle: Bundle.main), forCellReuseIdentifier: "cell")
 
         let searchViewController = UISearchController(searchResultsController: nil)
         navigationItem.searchController = searchViewController
@@ -61,8 +62,8 @@ extension SearchMoviesViewController: UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(
-            withIdentifier: "TestCell", for: indexPath
-            ) as? CellMovies
+            withIdentifier: "Cell", for: indexPath
+            ) as? CellMoviesXIB
         let movie = self.searchArray[indexPath.row]
         cell?.configure(movie: movie)
         return cell ?? UITableViewCell()
@@ -72,5 +73,19 @@ extension SearchMoviesViewController: UITableViewDataSource {
 
 extension SearchMoviesViewController: UISearchBarDelegate {
 
-
+//    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+//
+//        let searchAPI = """
+//        https://api.themoviedb.org/3/search/movie?api_key=4cb1eeab94f45affe2536f2c684a5c9e&query=\(searchText)
+//        """
+//        manager.getMovie(urlString: searchAPI, completion: { [weak self] getPopularMoviesResponse in
+//            guard let self = self else {
+//                return
+//            }
+//            DispatchQueue.main.async {
+//                self.searchArray = getPopularMoviesResponse?.results ?? []
+//                self.tableView?.reloadData()
+//            }
+//        })
+//    }
 }
